♻️ Refactor: 인증샷 관련 컴포저블을 design-system으로 이동
