+++ core/design-system/src/main/res/values/strings.xml
@@ -49,6 +49,10 @@
     <string name="comment_text_field_placeholder">코멘트추가</string>
     <string name="comment_condition_guide">5글자로 코멘트를 남길 수 있어요</string>
 
+    <!-- 인증샷 화면 -->
+    <string name="partner_not_task_certification">%s님은\n아직…</string>
+    <string name="partner_sting">찌르기</string>
+
     <!-- 목표 생성, 편집 화면 -->
     <string name="goal_editor_title">목표 직접 만들기</string>
     <string name="goal_editor_title_edit">목표 수정하기</string>
