+++ feature/task-certification/src/main/res/values/strings.xml
@@ -9,9 +9,7 @@
     <string name="comment_error_message">코멘트는 5글자로 입력해주세요!</string>
 
     <!-- 인증샷 보기 화면 !-->
-    <string name="task_certification_detail_partner_not_task_certification">%s님은\n아직…</string>
     <string name="task_certification_detail_fetch_photolog_fail">인증샷 조회에 실패했습니다.</string>
-    <string name="task_certification_detail_partner_sting">찌르기</string>
 
     <!-- 인증샷 수정 화면 !-->
     <string name="task_certification_editor_retake">다시 찍기</string>
